Local storage is functional. Testing for most paths is complete.
Need to think through any edge cases. Also the ids are being generated again on a load since the objects get assigned ids via a static method on creation. This should be fine since the storage is then updated to reflect these ids but more thought/testing would be good.