Archive list item styling and sorting completed.
In the future I'd like to re-incorporate an expanded view for archived items, similar to edit for normal items. Users may need to refer back to archived items and see the details/description info.
Also think about unarchiving? Do we have a use case for this? Why would a user need to do this?